abstracted out checkbox view to be inherited
